Move all setup stuff into a separate file
