+++ telepresence/outbound/__init__.py
@@ -11,19 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from telepresence.outbound.container import setup_container
-from telepresence.outbound.local import setup_inject, setup_vpn
-from telepresence.runner import Runner
 
+from telepresence.outbound.setup import setup
 
-def setup(runner: Runner, args):
-    if args.method == "inject-tcp":
-        return setup_inject(runner, args)
-
-    if args.method == "vpn-tcp":
-        return setup_vpn(runner, args)
-
-    if args.method == "container":
-        return setup_container(runner, args)
-
-    assert False, args.method
+__all__ = "setup",
